+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.5.6
 
 - Use group_by dependency. #71
+- Use library to create export button. #73